feat: Add support for more Linux distributions
This commit enhances the OS detection in both the shell script and the C
program to support a wider range of Linux distributions, including
various Arch-based and Debian-based systems.

The following distributions have been added:

- Arch-based: cachyos, garuda, artix, arco, steamos, blackarch
- Debian-based: kali, deepin, devuan, mx, pop

This change resolves the issue where CachyOS was not recognized as a
supported operating system.

Fixes #6